MAJ navbar/hiérarchie dossier et fichier + création page et slug
