+++ pages/index.tsx
@@ -30,7 +30,6 @@ export default function Home({ children }): any {
 
         <h1>Toutes les solutions pour apprendre à votre rythme</h1>
         <div>
-          <div>Welcome{user.name}</div>
           <Link href="/onlineStore" passHref={true}>
             <Image
               width={200}
